refactor: 가이드 모드 폴더 구조 수정
