+++ Frontend/src/components/guide-mode/GuideIntro/GuideIndicator/GuideIndicator.style.tsx
