+++ Frontend/src/components/guide-mode/GuideIntro/GuideMain/GuideMain.style.tsx
