+++ Frontend/src/components/guide-mode/GuideIntro/GuideMainComplete/GuideMainComplete.style.tsx
