+++ Frontend/src/components/guide-mode/GuideIntro/GuideMainComplete/GuideMainComplete.tsx
