+++ Frontend/src/components/guide-mode/GuideIntro/GuideMainTag/GuideMainTag.style.tsx
